+++ examples/clock.js
@@ -2,8 +2,7 @@ import { Canvas, Colors } from "../index.js";
 
 const canvas = new Canvas("clock", 600, 600);
 canvas.loadFont("roboto", "assets/Roboto-Regular.ttf")
-function drawHands() {
-    const date = new Date();
+function drawHands(date) {
     const seconds = date.getSeconds() / 60 * Math.PI * 2 - Math.PI / 2;
     const minutes = date.getMinutes() / 60 * Math.PI * 2 - Math.PI / 2;
     const hours = date.getHours() / 12 * Math.PI * 2 - Math.PI / 2;
@@ -11,7 +10,9 @@ function drawHands() {
     canvas.drawLine(Colors.AQUAMARINE, canvas.CENTER, Canvas.convertPolarCoords(canvas.CENTER, seconds, 170));
     canvas.drawLine(Colors.AQUAMARINE, canvas.CENTER, Canvas.convertPolarCoords(canvas.CENTER, minutes, 150));
     canvas.drawLine(Colors.AQUAMARINE, canvas.CENTER, Canvas.convertPolarCoords(canvas.CENTER, hours, 130));
+}
 
+function drawText(date) {
     let pos = { x: canvas.CENTER.x, y: canvas.CENTER.y };
     const width = 100;
     const height = 50;
@@ -22,6 +23,8 @@ function drawHands() {
 }
 
 canvas.loop(() => {
+    const date = new Date();
     canvas.drawArc(Colors.WHITE, canvas.CENTER, 200, 0, Math.PI * 2);
-    drawHands();
-})
+    drawText(date);
+    drawHands(date);
+});