+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # Changelog
 
+## 1.3.4
+
+- **`[Added]`** close window and end loop
+- **`[Fixed]`** detach
+
 ## 1.3.3
 
-- **`[Fixed]`** optimizations for path drawing
-- **`[Fixed]`** memory leak
-- **`[Fixed]`** fixed a problem with RGB888 format
 - **`[Added]`** move layer
 - **`[Added]`** attach buffer to video memory
 - **`[Added]`** PixelFormats class
 - **`[Changed]`** get layers returns the layer in order of appearance
+- **`[Fixed]`** optimizations for path drawing
+- **`[Fixed]`** memory leak
+- **`[Fixed]`** fixed a problem with RGB888 format
 
 ## 1.3.2